# docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: Wissenpos Restaurant Management Api
  description: Apis for the wissenspos application
  version: 1.0.0
servers:
- url: http://localhost:5000/api/v1
  description: Local development server
- url: https://wissenspos.herokuapp.com/api/v1
  description: Development Server
paths:
...
  "/accounts":
    get:
      tags:
      - accounts
      summary: Get a list of all accounts
      description: Restricted to admin users
      operationId: getallaccounts
      security:
      - bearerAuth: []
      responses:
        '200':
          description: An array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    firstName:
                      type: string
                      example: sunday
                    lastName:
                      type: string
                      example: oliseh
                    email:
                      type: string
                      example: sunday@oliseh.com
                    role:
                      type: string
                      example: customer
                    phoneNumber:
                      type: string
                      example: '0714382366'
                    created_at:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    updated_at:
                      type: string
                      example: '2020-05-08T03:11:21.553Z'
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
  "/accounts/{id}":
    parameters:
    - in: path
      name: id
      description: Account id
      required: true
      example: 1
      schema:
        type: integer
    get:
      tags:
      - accounts
      summary: Get a single account by id
      description: Admin users can access any account, regular users are restricted
        to their own account
      operationId: getaccountbyid
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Details of the specified account
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  firstName:
                    type: string
                    example: sunday
                  lastName:
                    type: string
                    example: oliseh
                  email:
                    type: string
                    example: sunday@oliseh.com
                  phoneNumber:
                    type: string
                    example: '071234569'
                  created_at:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  updated_at:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  role:
                    type: string
                    example: customer
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '404':
          "$ref": "#/components/responses/NotFoundError"
    put:
      tags:
      - accounts
      summary: Update an account
      description: Admin users can update any account, regular users are restricted
        to their own account and soon they wont be able to update role
      operationId: updateaccount
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: michael
                lastName:
                  type: string
                  example: smith
                email:
                  type: string
                  example: michael@smith.com
                password:
                  type: string
                  example: pass12345678
                confirmPassword:
                  type: string
                  example: pass12345678
                role:
                  type: string
                  enum:
                  - admin
                  - owner
                  - staff
                  - customer
      responses:
        '200':
          description: Account updated successfully. The details of the updated user
            are returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 51
                  firstName:
                    type: string
                    example: michael
                  lastName:
                    type: string
                    example: smith
                  email:
                    type: string
                    example: michael@smith.com
                  role:
                    type: string
                    example: staff
                  created_at:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  updated_at:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '404':
          "$ref": "#/components/responses/NotFoundError"
    delete:
      tags:
      - accounts
      summary: Delete an account
      description: Admin users can delete any account, regular users are restricted
        to their own account.
      operationId: deleteaccount
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Account deleted successfully
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '404':
          "$ref": "#/components/responses/NotFoundError"
...
  "/company/{company_id}/accounts":
    parameters:
    - in: path
      name: company_id
      description: Company ID
      required: true
      schema:
        type: integer
    get:
      tags:
      - company
      summary: Get accounts associated with a company
      description: Restricted to owners
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Array containing user objects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    company_id:
                      type: integer
                      example: 2
                    staff_id:
                      type: integer
                      example: 3
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
...
  "/company/{company_id}/order":
    parameters:
    - in: path
      name: company_id
      description: Company ID
      required: true
      schema:
        type: integer
    post:
      tags:
      - order
      summary: Add a product/item to cart
      description: add item
      operationId: addtocart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cart_id:
                  type: string
                  example: ADHIUYR7389F37T
                product_id:
                  type: integer
                  example: 1
                quantity:
                  type: integer
                  example: 10
      responses:
        '200':
          description: Item added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Item 1 was added successfully
        '500':
          description: cart_id is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: cart_id is required
  "/company/{company_id}/order/{cart_id}":
    parameters:
    - in: path
      name: company_id
      description: Company ID
      required: true
      schema:
        type: integer
    - in: path
      name: cart_id
      description: Cart Id
      required: true
      schema:
        type: string
    get:
      tags:
      - order
      summary: Get cart and items
      description: Fetch cart and the products i the cart
      operationId: getcartbyid
      responses:
        '200':
          description: Cart was fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cart_id:
                    type: string
                    example: ASHIEOHNDUUHUNIE
                  order_status:
                    type: string
                    example: New
                  purchase_status:
                    type: string
                    example: unpaid
                  subtotal:
                    type: number
                    example: 0
                  company_id:
                    type: integer
                    example: 1
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        order_id:
                          type: integer
                          example: 1
                        item_id:
                          type: integer
                          example: 2
                        name:
                          type: string
                          example: Beef Matumbo
                        price:
                          type: number
                          example: 250
                        image_url:
                          type: string
                          example: https://i.pinimg.com/736x/bf/6d/ff/bf6dff9d4f719e650beb1488208ba39d.jpg
                        quantity:
                          type: integer
                          example: 100
        '404':
          "$ref": "#/components/responses/NotFoundError"
  /orders/my-orders:
    get:
      tags:
        - "order"
      parameters:
        - in: query
          name: queryParams
          schema:
            type: object
            properties:
              nextPage:
                type: string
                example: " "
              limit:
                type: integer
                example: 30
                
      summary: Get a list of all your orders
      description: Has pagination and can accept query parmas to control the items fetched
      operationId: myorders
      responses:
        "200":
          description: obj containing order results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          example: 1
                        customer_id:
                          type: number
                          example: 1
                        company_id:
                          type: number
                          example: 1
                        order_status:
                          type: string
                          example: "Complete"
                        purchase_status:
                          type: string
                          example: "paid"
                        subtotal:
                          type: number
                          example: 898989.78
                        created_at:
                          type: string
                          example: "2020-05-08T03:11:21.553Z"
                        items:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: number
                                example: 1
                              order_id:
                                type: number
                                example: 1
                              item_id:
                                type: number
                                example: 1
                              quantity:
                                type: number
                                example: 45
                        pageInfo:
                          type: object
                          properties:
                            total:
                              type: integer
                              example: 30
                            remaining:
                              type: integer
                              example: 20
                            hasMore:
                              type: boolean
                              example: true
                            hasPrevious:
                              type: boolean
                              example: true
                            hasNext:
                              type: boolean
                              example: true
                            next:
                              type: string
                              example: "KGRhdGUpMjAyMS0wMi0wNFQwMDo1ODoyOS4xODNa"
                            previous:
                              type: string
                              example: "KGRhdGUpMjAyMS0wMi0wNFQwMDo1ODoyOS4xODNat"  
        "404":
          "$ref": "#/components/responses/NotFoundError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access
        to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Not Found
...